+++ R/cat2cat.R
@@ -1,52 +1,66 @@
-#' Automatic mapping of a categorical variable in a panel dataset according to a new encoding
-#' @description This function is built to work for two time points at once.
-#' Thus for more periods some recursion will be needed.
-#' The \code{prune_c2c} might be needed when we have many interactions to limit growing number of replications.
-#' This function might seems to be a complex at the first glance though it is built to offer a wide range of applications for complex tasks.
-#' @param data `named list` with fields `old`, `new`, `cat_var` (or `cat_var_old` and `cat_var_new`), `time_var` and optional `id_var`,`multiplier_var`,`freq_df`.
-#' @param mappings `named list` with 2 fields `trans` and `direction`.
-#' @param ml `named list` with up to 5 fields `data`, `cat_var`, `method`, `features` and optional `args`.
+#' Automatic mapping in a panel dataset
+#' @description
+#' The objective is to unify an inconsistently coded categorical variable in a panel
+#' dataset according to a mapping (transition) table.
+#' The mapping (transition) table is the core element of the process.
+#' There are three arguments `data`, `mappings`, and `ml`. Each
+#' of these arguments is of a `list` type, wherein the
+#' `ml` argument is optional. Arguments are separated to
+#' identify the core elements of the `cat2cat` procedure. Although this function seems
+#' complex initially, it is built to offer a wide range of
+#' applications for complex tasks. The function contains
+#' many validation checks to prevent incorrect usage.
+#' The function has to be applied iteratively for each two neighboring periods of a panel dataset.
+#' The \code{prune_c2c} function could be needed to limit growing number of replications.
+#' @param data `named list` with fields `old`, `new`, `cat_var` (or `cat_var_old` and `cat_var_new`), `time_var` and optional `id_var`,`multiplier_var`.
+#' @param mappings `named list` with 3 fields `trans`, `direction` and optional `freq_df`.
+#' @param ml `named list` (optional) with up to 5 fields `data`, `cat_var`, `method`, `features` and optional `args`.
 #' @details
 #' data args
 #' \itemize{
 #'  \item{"old"}{ data.frame older time point in a panel}
 #'  \item{"new"} { data.frame more recent time point in a panel}
-#'  \item{"cat_var"}{ character name of the categorical variable.}
-#'  \item{"cat_var_old"}{ optional character name of the categorical variable in the older time point. Default `cat_var`.}
-#'  \item{"cat_var_new"}{ optional character name of the categorical variable in the newer time point. Default `cat_var`.}
-#'  \item{"time_var"}{ character name of the time variable.}
-#'  \item{"id_var"}{ optional character name of the unique identifier variable - if this is specified then for subjects observe in both periods the direct mapping is applied.}
-#'  \item{"multiplier_var"}{ optional character name of the multiplier variable - number of replication needed to reproduce the population}
-#'  \item{"freqs_df"}{ optional - data.frame with 2 columns where first one is category name and second counts.
-#'  It is optional nevertheless will be very often needed, as give more control.
-#'  It will be used to assess the probabilities. The multiplier variable is omit so sb has to apply it in this table.}
+#'  \item{"time_var"}{ character(1) name of the time variable.}
+#'  \item{"cat_var"}{ character(1) name of the categorical variable.}
+#'  \item{"cat_var_old"}{ Optional character(1) name of the categorical variable in the older time point. Default `cat_var`.}
+#'  \item{"cat_var_new"}{ Optional character(1) name of the categorical variable in the newer time point. Default `cat_var`.}
+#'  \item{"id_var"}{ Optional character(1) name of the unique identifier variable - if this is specified then for subjects observe in both periods the direct mapping is applied.}
+#'  \item{"multiplier_var"}{ Optional character(1) name of the multiplier variable - number of replication needed to reproduce the population}
+#'  \item{"freqs_df"}{ For backward compaibility check the definition in the description. of the mappings argument }
 #' }
 #' mappings args
 #' \itemize{
-#'  \item{"trans"}{ data.frame with 2 columns - transition table - all categories for cat_var in old and new datasets have to be included.
+#'  \item{"trans"}{ data.frame with 2 columns - mapping (transition) table - all categories for cat_var in old and new datasets have to be included.
 #'   First column contains an old encoding and second a new one.
-#'   The transition table should to have a candidate for each category from the targeted for an update period.
+#'   The mapping (transition) table should to have a candidate for each category from the targeted for an update period.
 #' }
-#'  \item{"direction"}{ character direction - "backward" or "forward"}
+#'  \item{"direction"}{ character(1) direction - "backward" or "forward"}
+#'  \item{"freqs_df"}{ Optional - data.frame with 2 columns where first one is category name (base period) and second counts.
+#'  If It is not provided then is assessed automatically.
+#'  Artificial counts for each variable level in the base period.
+#'  It is optional nevertheless will be often needed, as gives more control.
+#'  It will be used to assess the probabilities. The multiplier variable is omitted so sb has to apply it in this table.}
 #' }
-#' optional ml args
+#' Optional ml args
 #' \itemize{
 #'  \item{"data"}{ data.frame - dataset with features and the `cat_var`.}
-#'  \item{"cat_var"}{ character - the dependent variable name.}
+#'  \item{"cat_var"}{ character(1) - the dependent variable name.}
 #'  \item{"method"}{ character vector - one or a few from "knn", "rf" and "lda" methods - "knn" k-NearestNeighbors, "lda" Linear Discrimination Analysis, "rf" Random Forest }
 #'  \item{"features"}{ character vector of features names where all have to be numeric or logical}
 #'  \item{"args"}{ optional - list parameters: knn: k ; rf: ntree  }
 #' }
-#' @return named list with 2 fields old an new - 2 data.frames.
+#' @return `named list` with 2 fields old and new - 2 data.frames.
 #' There will be added additional columns like index_c2c, g_new_c2c, wei_freq_c2c, rep_c2c, wei_(ml method name)_c2c.
-#' Additional columns will be informative only for a one data.frame as we always make a changes to one direction.
+#' Additional columns will be informative only for a one data.frame as we always make the changes to one direction.
 #' @importFrom stats predict complete.cases setNames
 #' @importFrom MASS lda
-#' @details Without ml section only simple frequencies are assessed.
+#' @details
+#' Without ml section only simple frequencies are assessed.
 #' When ml model is broken then weights from simple frequencies are taken.
 #' `knn` method is recommended for smaller datasets.
 #' @note
-#' The transition table should to have a candidate for each category from the targeted for an update period.
+#' `trans` arg columns and the `cat_var` column have to be of the same type.
+#' The mapping (transition) table should to have a candidate for each category from the targeted for an update period.
 #' The observation from targeted for an updated period without a matched category from base period is removed.
 #' If you want to leave NA values add `c(NA, NA)` row to the `trans` table.
 #' Please check the vignette for more information.
@@ -82,7 +96,7 @@
 #'   data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
 #'   mappings = list(trans = trans, direction = "backward"),
 #'   ml = list(
-#'     data = occup_old,
+#'     data = occup_small[occup_small$year >= 2010, ],
 #'     cat_var = "code",
 #'     method = "knn",
 #'     features = c("age", "sex", "edu", "exp", "parttime", "salary"),
@@ -94,11 +108,11 @@
 cat2cat <- function(data = list(
                       old = NULL,
                       new = NULL,
+                      time_var = NULL,
                       cat_var = NULL,
                       cat_var_old = NULL,
                       cat_var_new = NULL,
                       id_var = NULL,
-                      time_var = NULL,
                       multiplier_var = NULL,
                       freqs_df = NULL
                     ),
@@ -126,6 +140,8 @@ cat2cat <- function(data = list(
     is.null(data$multiplier_var) ||
       (data$multiplier_var %in% colnames(data$new) && data$multiplier_var %in% colnames(data$old))
   )
+
+  # For backward compatibility
   stopifnot(is.null(data$freqs_df) || (is.data.frame(data$freqs_df) && ncol(data$freqs_df) == 2))
   stopifnot((length(unique(data$old[[data$time_var]])) == 1) && (length(unique(data$new[[data$time_var]])) == 1))
   stopifnot(is.null(data$id_var) || ((data$id_var %in% colnames(data$old)) &&
@@ -139,77 +155,51 @@ cat2cat <- function(data = list(
   stopifnot(isTRUE(mappings$direction %in% c("forward", "backward")))
   stopifnot(is.data.frame(mappings$trans) && ncol(mappings$trans) == 2)
 
-  is_direct_match <- !is.null(data$id_var)
   mapps <- get_mappings(mappings$trans)
 
-  if (is_direct_match) {
-    id_inner <- intersect(data$old[[data$id_var]], data$new[[data$id_var]])
-    tos <- merge(data$old[, c(data$id_var, data$cat_var_old)], data$new[, c(data$id_var, data$cat_var_new)], by = data$id_var)
-    colnames(tos) <- c(data$id_var, "cat_old", "cat_new")
-    if (mappings$direction == "forward") {
-      id_outer <- setdiff(data$new[[data$id_var]], data$old[[data$id_var]])
-      tos_df <- tos[, c(data$id_var, "cat_old")]
-    } else if (mappings$direction == "backward") {
-      id_outer <- setdiff(data$old[[data$id_var]], data$new[[data$id_var]])
-      tos_df <- tos[, c(data$id_var, "cat_new")]
-    }
-    colnames(tos_df) <- c("id", "cat")
+  if (mappings$direction == "forward") {
+    base_name <- "old"
+    target_name <- "new"
+  } else if (mappings$direction == "backward") {
+    base_name <- "new"
+    target_name <- "old"
   }
 
-  if (mappings$direction == "forward") {
-    cat_var_base <- data$cat_var_old
-    cat_var_target <- data$cat_var_new
-    cat_base_year <- data$old
-    cat_target_year <- if (is_direct_match) data$new[data$new[[data$id_var]] %in% id_outer, ] else data$new
-    cat_mid <- if (is_direct_match) data$new[data$new[[data$id_var]] %in% id_inner, ] else NULL
-    mapp <- mapps$to_old
-    res_ord <- c(1, 2)
-  } else if (mappings$direction == "backward") {
-    cat_var_base <- data$cat_var_new
-    cat_var_target <- data$cat_var_old
-    cat_base_year <- data$new
-    cat_target_year <- if (is_direct_match) data$old[data$old[[data$id_var]] %in% id_outer, ] else data$old
-    cat_mid <- if (is_direct_match) data$old[data$old[[data$id_var]] %in% id_inner, ] else NULL
-    mapp <- mapps$to_new
-    res_ord <- c(2, 1)
+  cat_base_year <- data[[base_name]]
+  cat_target_year <- data[[target_name]]
+  cat_var_base <- data[[paste0("cat_var_", base_name)]]
+  cat_var_target <- data[[paste0("cat_var_", target_name)]]
+  cat_mid <- NULL # direct match
+
+  is_direct_match <- !is.null(data$id_var)
+  if (is_direct_match) {
+    id_inner <- intersect(data$old[[data$id_var]], data$new[[data$id_var]])
+    id_outer <- setdiff(data[[target_name]][[data$id_var]], data[[base_name]][[data$id_var]])
+
+    cat_target_year <- data[[target_name]][data[[target_name]][[data$id_var]] %in% id_outer, ]
+    cat_mid <- dummy_c2c(data[[target_name]][data[[target_name]][[data$id_var]] %in% id_inner, ], cat_var_base)
+
+    tos <- merge(data$old[, c(data$id_var, data$cat_var_old)], data$new[, c(data$id_var, data$cat_var_new)], by = data$id_var)
+    colnames(tos) <- c("id", "cat_old", "cat_new")
+    cat_mid$g_new_c2c <- tos[[paste0("cat_", base_name)]][match(cat_mid[[data$id_var]], tos$id)]
   }
 
   cats_base <- cat_base_year[[cat_var_base]]
   cats_target <- cat_target_year[[cat_var_target]]
 
-  if (cats_target_diff_len <- length(cats_target_diff <- setdiff(unique(cats_target), names(mapp)))) {
-    warning(
-      paste(
-        sprintf("trans table does not cover some levels for a target period, so the result will not contain all original observations. Lacking %s levels: ",
-                cats_target_diff_len),
-        paste(head(cats_target_diff, 10), collapse = ", "),
-        if (cats_target_diff_len >= 10) "..." else NULL)
-    )
-  }
+  mapp <- mapps[[paste0("to_", base_name)]]
 
-  if (is_direct_match) {
-    cat_mid <- dummy_c2c(cat_mid, cat_var_base)
-    cat_mid$g_new_c2c <- tos_df$cat[match(cat_mid[[data$id_var]], tos_df$id)]
-  }
+  validate_cover_cats(unique(cats_target), mapp)
 
-  fre <- if (is.data.frame(data$freqs_df)) {
-    data$freqs_df
-  } else if ("wei_freq_c2c" %in% colnames(cat_base_year)) {
-    stats::aggregate(
-      if (!is.null(data$multiplier_var)) cat_base_year[["wei_freq_c2c"]] * cat_base_year[[data$multiplier_var]] else cat_base_year[["wei_freq_c2c"]],
-      list(g = cat_base_year[[cat_var_base]]),
-      function(x) round(sum(x, na.rm = TRUE))
-    )
-  } else {
-    if (!is.null(data$multiplier_var)) {
-      get_freqs(cats_base, cat_base_year[[data$multiplier_var]])
-    } else {
-      get_freqs(cats_base)
-    }
-  }
+  # frequencies
+  fre <- resolve_frequencies(
+    cat_base_year = cat_base_year, cat_var_base = cat_var_base,
+    freqs_df = data$freqs_df, multiplier_var = data$multiplier_var
+  )
+  # frequencies per category
   freqs_list <- cat_apply_freq(mapp, fre)
 
-  # Replicate and apply freq probabilities
+  # Replicate and apply freq probabilities, target period
   g_vec <- mapp[match(cats_target, names(mapp))]
   rep_vec <- unname(lengths(g_vec))
   wei_vec <- freqs_list[match(cats_target, names(freqs_list))]
@@ -245,20 +235,21 @@ cat2cat <- function(data = list(
 
     cat_target_rep <- ml_results$target_data
   }
-
+  # bind direct match if applied
   cat_target_rep_f <- rbind(cat_target_rep, cat_mid)
 
-  res <- list(cat_base_year, cat_target_rep_f)[res_ord]
-  names(res) <- c("old", "new")
+  res <- list()
+  res[[base_name]] <- cat_base_year
+  res[[target_name]] <- cat_target_rep_f
   res
 }
 
-#' The internal function used in the cat2cat ones
+#' The internal function used in the cat2cat one
 #' @description apply the ml models to the cat2cat data
 #' @param ml `list` the same `ml` argument as provided to `cat2cat` function.
 #' @param mapp `list` a mapping table
 #' @param target_data `data.frame`
-#' @param cat_var_target `character` name of the categorical variable in the target period.
+#' @param cat_var_target `character(1)` name of the categorical variable in the target period.
 #' @keywords internal
 cat2cat_ml <- function(ml, mapp, target_data, cat_var_target) {
   stopifnot(all(c("method", "features") %in% names(ml)))
@@ -295,7 +286,7 @@ cat2cat_ml <- function(ml, mapp, target_data, cat_var_target) {
           length(unique(target_data_cat$g_new_c2c)) > 1 &&
             length(udc) >= 1 &&
             nrow(target_data_cat) > 0 &&
-            any(target_data_cat$g_new_c2c %in% names(cat_ml_year_g))
+            any(unique(target_data_cat$g_new_c2c) %in% names(cat_ml_year_g))
         ) {
           base_ml <- target_data_cat[!duplicated(target_data_cat[["index_c2c"]]), c("index_c2c", features)]
           cc <- complete.cases(base_ml[, features])
@@ -359,4 +350,48 @@ cat2cat_ml <- function(ml, mapp, target_data, cat_var_target) {
   target_data <- target_data[order(target_data[["index_c2c"]]), ]
 
   list(target_data = target_data)
+}
+
+#' Validate if the trans table contains all proper mappings
+#' @param cats_target vector of unique target period categories
+#' @param mapp transition (mapping) table process with `get_mappings`, the "to_base" direction is taken.
+#' @keywords internal
+validate_cover_cats <- function(u_cats_target, mapp) {
+  if (cats_target_diff_len <- length(cats_target_diff <- setdiff(u_cats_target, names(mapp)))) {
+    warning(
+      paste(
+        sprintf(
+          paste0(
+            "trans table does not cover some levels for a target period, ",
+            "so the result will not contain all original observations. Lacking %s levels: "
+          ),
+          cats_target_diff_len
+        ),
+        paste(head(cats_target_diff, 10), collapse = ", "),
+        if (cats_target_diff_len >= 10) "..." else NULL
+      )
+    )
+  }
+}
+
+#' Resolve the frequencies
+#' @param cat_base_year `data.frame` a base period dataset.
+#' @param cat_var_base `character(1)` name of the base period categorical variable.
+#' @param freqs_df `data.frame` with 2 columns, the first one with category and second with counts.
+#' @param multiplier_var `character(1)` name of the multiplier variable.
+#' @keywords internal
+resolve_frequencies <- function(cat_base_year, cat_var_base, freqs_df, multiplier_var) {
+  cats_base <- cat_base_year[[cat_var_base]]
+  freqs <- if (is.data.frame(freqs_df)) {
+    freqs_df
+  } else if ("wei_freq_c2c" %in% colnames(cat_base_year)) {
+    stats::aggregate(
+      if (!is.null(multiplier_var)) cat_base_year[["wei_freq_c2c"]] * cat_base_year[[multiplier_var]] else cat_base_year[["wei_freq_c2c"]],
+      list(g = cats_base),
+      function(x) round(sum(x, na.rm = TRUE))
+    )
+  } else {
+    get_freqs(cats_base, if (!is.null(multiplier_var)) cat_base_year[[multiplier_var]])
+  }
+  freqs
 }