+++ R/mapping.R
@@ -1,10 +1,10 @@
-#' Transforming a transition table with mappings to two associative lists
+#' Transforming a mapping (transition) table to two associative lists
 #' @description to rearrange the one classification encoding into another, an associative list that maps keys to values is used.
 #' More precisely, an association list is used which is a linked list in which each list element consists of a key and value or values.
 #' An association list where unique categories codes are keys and matching categories from next or previous time point are values.
-#' A transition table is used to build such associative lists.
-#' @param x data.frame or matrix - transition table with 2 columns where first column is assumed to be the older encoding.
-#' @details the named list will be a more efficient solution than hash map as we are not expecting more than a few thousand keys.
+#' A mapping (transition) table is used to build such associative lists.
+#' @param x `data.frame` or `matrix` - mapping (transition) table with 2 columns where first column is assumed to be the older encoding.
+#' @details the `named list` will be almost equally efficient solution as the hash map (`environment`) as we are not expecting more than a few thousand keys and R has an in-build multi-element subset operator for lists.
 #' @return a list with 2 named lists `to_old` and `to_new`.
 #' @examples
 #' data(trans)
@@ -14,6 +14,7 @@
 #' mappings$to_new[1:4]
 #' @export
 get_mappings <- function(x = data.frame()) {
+  stopifnot(is.data.frame(x) || is.matrix(x))
   stopifnot(ncol(x) == 2)
 
   x <- as.matrix(x)
@@ -47,13 +48,46 @@ get_mappings <- function(x = data.frame()) {
   list(to_old = to_old, to_new = to_new)
 }
 
-#' Applying frequencies to the object returned by get_mappings function
-#' @description applying frequencies to the object returned by get_mappings.
-#' We will get a symmetric object to returned one by get_mappings function, nevertheless categories are replaced with frequencies.
+#' Getting frequencies from a vector with an optional multiplier
+#' @description getting frequencies for a vector with an optional multiplier.
+#' @param x `vector` categorical variable to summarize.
+#' @param multiplier `numeric` vector how many times to repeat certain value, additional weights.
+#' @return `data.frame` with two columns `input` `Freq`
+#' @note without multiplier variable it is a basic `table` function wrapped with the `as.data.frame` function.
+#' The `table` function is used with the `useNA = "ifany"` argument.
+#' @export
+#' @examples
+#' data(occup)
+#'
+#' head(get_freqs(occup$code[occup$year == "2008"]))
+#' head(get_freqs(occup$code[occup$year == "2010"]))
+#'
+#' head(get_freqs(occup$code[occup$year == "2008"], occup$multiplier[occup$year == "2008"]))
+#' head(get_freqs(occup$code[occup$year == "2010"], occup$multiplier[occup$year == "2010"]))
+get_freqs <- function(x, multiplier = NULL) {
+  stopifnot(is.vector(x))
+  stopifnot(is.null(multiplier) || length(x) == length(multiplier))
+
+  input <- if (!is.null(multiplier)) {
+    rep(x, times = as.numeric(multiplier))
+  } else {
+    x
+  }
+  res <- as.data.frame(table(input, useNA = "ifany"), stringsAsFactors = F)
+  res
+}
+
+
+#' Applying frequencies to the object returned by the `get_mappings` function
+#' @description applying frequencies to the object returned by the `get_mappings` function.
+#' We will get a symmetric object to the one returned by the `get_mappings` function, nevertheless categories are replaced with frequencies.
 #' Frequencies for each category/key are sum to 1, so could be interpreted as probabilities.
-#' @param to_x list object returned by get_mappings.
-#' @param freqs vector object returned by get_freqs.
-#' @return a list with 2 named lists `to_old` and `to_new`.
+#' @param to_x `list` object returned by `get_mappings`.
+#' @param freqs `data.frame` object like the one returned by the `get_freqs` function.
+#' @return a `list` with a structure like `to_x` object but with probabilities for each category.
+#' @note
+#' `freqs` arg first column (keys) and the to_x arg values have to be of the same type.
+#' The uniform distribution (outcomes are equally likely) is assumed for no match for all possible categories.
 #' @examples
 #' data(trans)
 #' data(occup)
@@ -81,6 +115,7 @@ get_mappings <- function(x = data.frame()) {
 #' head(data.frame(I(mappings$to_new), I(mapp_p)))
 #' @export
 cat_apply_freq <- function(to_x, freqs) {
+  stopifnot(is.list(to_x))
   stopifnot(ncol(freqs) == 2)
   res <- lapply(
     to_x,
@@ -88,44 +123,15 @@ cat_apply_freq <- function(to_x, freqs) {
       alls <- freqs[, 2, drop = TRUE][match(x, freqs[, 1, drop = TRUE])]
       ff <- alls / sum(alls, na.rm = T)
       # NA to 0
-      ifelse(is.na(ff) | is.nan(ff), 0, ff)
+      ff[is.na(ff) | is.nan(ff)] <- 0
+      # all equal to zero so proportional probability
+      if (all(ff == 0)) {
+        ff_len <- length(ff)
+        ff <- rep(1 / ff_len, ff_len)
+      }
+      ff
     }
   )
-  # all equal to zero so proportional probability
-  res_out <- lapply(
-    res,
-    function(x) {
-      ifelse(rep(all(x == 0), length(x)), 1 / length(x), x)
-    }
-  )
-  names(res_out) <- names(to_x)
-  res_out
-}
-
-#' Getting frequencies from a `character` vector with an optional multiplier argument
-#' @description getting frequencies for a vector with an optional multiplier argument
-#' @param x character vector categorical variable to summarize.
-#' @param multiplier numeric vector how many times to repeat certain value, additional weights.
-#' @return data.frame with two columns `input` `Freq`
-#' @note without multiplier variable it is a basic `table` function wrapped with the `as.data.frame` function.
-#' The `table` function is used with the `useNA = "ifany"` argument.
-#' @export
-#' @examples
-#' data(occup)
-#'
-#' head(get_freqs(occup$code[occup$year == "2008"]))
-#' head(get_freqs(occup$code[occup$year == "2010"]))
-#'
-#' head(get_freqs(occup$code[occup$year == "2008"], occup$multiplier[occup$year == "2008"]))
-#' head(get_freqs(occup$code[occup$year == "2010"], occup$multiplier[occup$year == "2010"]))
-get_freqs <- function(x, multiplier = NULL) {
-  stopifnot(is.null(multiplier) || length(x) == length(multiplier))
-
-  input <- if (!is.null(multiplier)) {
-    rep(x, times = as.numeric(multiplier))
-  } else {
-    x
-  }
-  res <- as.data.frame(table(input, useNA = "ifany"), stringsAsFactors = F)
+  names(res) <- names(to_x)
   res
 }