+++ tests/testthat/test-cat2cat.R
@@ -51,19 +51,22 @@ expect_equal(
   (occup_2$old$wei_knn_c2c + occup_2$old$wei_freq_c2c) / 2
 )
 
-expect_warning({
-  cat2cat(
-    data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
-    mappings = list(trans = head(trans, -50), direction = "backward"),
-    ml = list(
-      data = occup_new,
-      cat_var = "code",
-      method = c("knn", "rf", "lda"),
-      features = c("age", "sex", "edu", "exp", "parttime", "salary"),
-      args = list(k = 10, ntree = 30)
+expect_warning(
+  {
+    cat2cat(
+      data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
+      mappings = list(trans = head(trans, -50), direction = "backward"),
+      ml = list(
+        data = occup_new,
+        cat_var = "code",
+        method = c("knn", "rf", "lda"),
+        features = c("age", "sex", "edu", "exp", "parttime", "salary"),
+        args = list(k = 10, ntree = 30)
+      )
     )
-  )
-}, "9321, 9312, 9311, 9331, 9313")
+  },
+  "9321, 9312, 9311, 9331, 9313"
+)
 
 ###############
 
@@ -133,14 +136,17 @@ expect_identical(nrow(occup_3b$old), nrow(occup_3c$old) + 1L)
 expect_true((all(occup_3c$old$wei_freq_c2c <= 1 & occup_3c$old$wei_freq_c2c >= 0)))
 expect_true((all(occup_3c$old$wei_knn_c2c <= 1 & occup_3c$old$wei_knn_c2c >= 0)))
 
-occup_4 <- expect_warning({
-  cat2cat(
-    data = list(
-      old = occup_old, new = occup_new, cat_var = "code", time_var = "year"
-    ),
-    mappings = list(trans = trans, direction = "forward")
-  )
-}, "741103, 712604, 732201, 818116, 732301, 816003, 741201, 713201, 818115, 815204")
+occup_4 <- expect_warning(
+  {
+    cat2cat(
+      data = list(
+        old = occup_old, new = occup_new, cat_var = "code", time_var = "year"
+      ),
+      mappings = list(trans = trans, direction = "forward")
+    )
+  },
+  "741103, 712604, 732201, 818116, 732301, 816003, 741201, 713201, 818115, 815204"
+)
 
 # not in trans table
 expect_equal(sum(occup_4$new$wei_freq_c2c) + sum(occup_new$code %in% setdiff(occup_new$code, trans$new)), nrow(occup_new))
@@ -150,8 +156,10 @@ occup_4b <- cat2cat(
   data = list(
     old = occup_old, new = occup_new, cat_var = "code", time_var = "year"
   ),
-  mappings = list(trans = rbind(trans, data.frame(old = "no_cat", new = setdiff(occup_new$code, trans$new))),
-                  direction = "forward")
+  mappings = list(
+    trans = rbind(trans, data.frame(old = "no_cat", new = setdiff(occup_new$code, trans$new))),
+    direction = "forward"
+  )
 )
 
 expect_identical(nrow(occup_4b$old), nrow(occup_4b$old))
@@ -164,7 +172,7 @@ data(verticals2)
 vert_old <- verticals2[verticals2$v_date == "2020-04-01", ]
 vert_new <- verticals2[verticals2$v_date == "2020-05-01", ]
 
-## get transitions table
+## get mapping (transition) table
 trans_v <- vert_old %>%
   inner_join(vert_new, by = "ean") %>%
   select(vertical.x, vertical.y) %>%
@@ -210,6 +218,8 @@ expect_true((all(verts2$old$wei_knn_c2c <= 1 & verts2$old$wei_knn_c2c >= 0)))
 expect_true((all(verts2$old$wei_rf_c2c <= 1 & verts2$old$wei_rf_c2c >= 0)))
 expect_true((all(verts2$old$wei_lda_c2c <= 1 & verts2$old$wei_lda_c2c >= 0)))
 
+expect_true(all(unique(verts2[["old"]][["g_new_c2c"]]) %in% trans_v[[2]]))
+
 verts3 <- cat2cat(
   data = list(old = vert_old, new = vert_new, id_var = "ean", cat_var = "vertical", time_var = "v_date"),
   mappings = list(trans = trans_v, direction = "forward"),
@@ -234,3 +244,5 @@ expect_equal(sum(verts3$new$wei_lda_c2c), nrow(vert_new))
 expect_true((all(verts3$new$wei_knn_c2c <= 1 & verts3$new$wei_knn_c2c >= 0)))
 expect_true((all(verts3$new$wei_rf_c2c <= 1 & verts3$new$wei_rf_c2c >= 0)))
 expect_true((all(verts3$new$wei_lda_c2c <= 1 & verts3$new$wei_lda_c2c >= 0)))
+
+expect_true(all(unique(verts3[["new"]][["g_new_c2c"]]) %in% trans_v[[1]]))